Rename lit-html.ts to core.ts
