+++ .gitignore
@@ -8,6 +8,9 @@
 /lit-html.d.ts
 /lit-html.js
 /lit-html.js.map
+/core.d.ts
+/core.js
+/core.js.map
 /test/**/*.d.ts
 /test/**/*.js
 /test/**/*.js.map
