+++ package.json
@@ -13,6 +13,9 @@
     "/lit-html.js",
     "/lit-html.js.map",
     "/lit-html.d.ts",
+    "/core.js",
+    "/core.js.map",
+    "/core.d.ts",
     "/lib/",
     "/src/",
     "!/src/test/"
@@ -22,7 +25,7 @@
     "gen-docs": "typedoc --readme none --mode modules --excludeNotExported --excludePrivate --exclude **/*_test.ts --out ./docs/api ./src",
     "test": "npm run build && npm run lint && wct --npm",
     "quicktest": "wct -l chrome -p --npm",
-    "checksize": "uglifyjs lit-html.js -mc --toplevel | gzip -9 | wc -c",
+    "checksize": "uglifyjs core.js -mc --toplevel | gzip -9 | wc -c",
     "format": "find src test | grep '\\.js$\\|\\.ts$' | xargs clang-format --style=file -i",
     "lint": "tslint --project ./",
     "prepare": "npm run build"
