+++ src/test/lib/lit-extended_test.ts
@@ -15,9 +15,8 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
+import {directive, html as htmlPlain} from '../../core.js';
 import {html, PropertyPart, render} from '../../lib/lit-extended.js';
-import {directive, html as htmlPlain} from '../../lit-html.js';
-
 import {stripExpressionDelimeters} from '../test-helpers.js';
 
 const assert = chai.assert;
