+++ src/test/lib/modify-template_test.ts
@@ -15,8 +15,8 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
+import {defaultTemplateFactory as templateFactory, html, render} from '../../core.js';
 import {insertNodeIntoTemplate, removeNodesFromTemplate} from '../../lib/modify-template.js';
-import {defaultTemplateFactory as templateFactory, html, render} from '../../lit-html.js';
 import {stripExpressionDelimeters} from '../test-helpers.js';
 
 const assert = chai.assert;
